pyproject.toml
[tool.poetry]
name = "to-inbox-api"
version = "1.0.0"
description = ""
authors = ["Andreas Backström <mozzo242@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.12"
Flask = "3.0.3"
python-dotenv = "1.0.1"
Flask-Cors = "4.0.1"
psycopg2 = "2.9.9"
flask-marshmallow = "1.2.1"
Flask-SQLAlchemy = "3.1.1"
marshmallow-sqlalchemy = "1.1.0"
Flask-Migrate = "^4.0.7"
gunicorn = "23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
